fix: show contrast warning across editor tabs
