+++ src/ui/settings/theme/theme-editor-quick.browser.test.tsx
@@ -39,18 +39,4 @@ describe("ThemeEditorQuick", () => {
     color.dispatchEvent(new Event("input", { bubbles: true }));
     expect(onTokenChange).toHaveBeenCalledWith("accent", "#abcdef");
   });
-
-  it("hides the contrast warning for a high-contrast draft", async () => {
-    await render(<ThemeEditorQuick draft={draftFrom("default")} onTokenChange={() => {}} />);
-    expect(document.body.textContent).not.toContain("below WCAG AA");
-  });
-
-  it("shows the contrast warning for a low-contrast draft", async () => {
-    const draft = draftFrom("default", {
-      tokens: { ...PRESET_BY_ID.get("default")?.tokens, bg: "#fefefe", text: "#ffffff" },
-    });
-    const screen = await render(<ThemeEditorQuick draft={draft} onTokenChange={() => {}} />);
-    await expect.element(screen.getByRole("alert")).toBeInTheDocument();
-    expect(document.body.textContent).toContain("below WCAG AA");
-  });
 });
